Fix drifting below Y=0
Vehicles now properly count as touching the ground when below Y=0.  This fixes not only drifting, but also Launch Gel and the ability for entities to hop inside of vehicles, as well as any other location where automobileOnGround is relevant.

This feels hacky. I don't know why this works. I don't know why it wasn't working to begin with. This probably will create some unexpected behavior elsewhere but we will simply have to wait and see on that one.

TODO: remove all the logging shit i commented out